## Separate the CUDA pipeline from the search algorithm

This PR keeps two things apart.

**Using the GPU**: copying frames to the device, spreading them over
several
CUDA streams, copying results back... This is the same no matter how
clusters
are found. It lives in `ClusterFinderCUDA`, which never refers to the
search code.

**Finding the clusters**: this lives in `clusterfinder_algo.cuh`, in
three
parts:

- `CenteredRunningPedestal` is the pedestal: how it is stored on the
GPU, and
  how it moves to and from the host
- `FixedWindow` the search algorithm
- `FixedWindow::Output` is what one frame writes out, and how to read it
back

@tomascelko A sparse finder is a basically a new search plus a new
`Output`.
Thepedestal and everything the pipeline does stay as they are. What a
search
has to provide is listed at the top of `clusterfinder_algo.cuh`, and
`require_algo`
checks a few of those at compile time (concepts would be cleaner, but
that
needs C++20).

One thing to know: `Output` uses the same number of bytes for every
frame,
because the host buffers are allocated before any frame runs. A search
that
writes a different amount per frame would either reserve room for the
worst
case, as the current one does, or that part would have to change...

+++ include/aare/ArrayExpr.hpp
@@ -29,7 +29,7 @@ class ArrayAdd : public ArrayExpr<ArrayAdd<A, B, Ndim>, Ndim> {
     ArrayAdd(const A &arr1, const B &arr2) : arr1_(arr1), arr2_(arr2) {
         assert(arr1.size() == arr2.size());
     }
-    auto operator[](int i) const { return arr1_[i] + arr2_[i]; }
+    auto operator[](size_t i) const { return arr1_[i] + arr2_[i]; }
     size_t size() const { return arr1_.size(); }
     std::array<ssize_t, Ndim> shape() const { return arr1_.shape(); }
 };
@@ -43,7 +43,7 @@ class ArraySub : public ArrayExpr<ArraySub<A, B, Ndim>, Ndim> {
     ArraySub(const A &arr1, const B &arr2) : arr1_(arr1), arr2_(arr2) {
         assert(arr1.size() == arr2.size());
     }
-    auto operator[](int i) const { return arr1_[i] - arr2_[i]; }
+    auto operator[](size_t i) const { return arr1_[i] - arr2_[i]; }
     size_t size() const { return arr1_.size(); }
     std::array<ssize_t, Ndim> shape() const { return arr1_.shape(); }
 };
@@ -57,7 +57,7 @@ class ArrayMul : public ArrayExpr<ArrayMul<A, B, Ndim>, Ndim> {
     ArrayMul(const A &arr1, const B &arr2) : arr1_(arr1), arr2_(arr2) {
         assert(arr1.size() == arr2.size());
     }
-    auto operator[](int i) const { return arr1_[i] * arr2_[i]; }
+    auto operator[](size_t i) const { return arr1_[i] * arr2_[i]; }
     size_t size() const { return arr1_.size(); }
     std::array<ssize_t, Ndim> shape() const { return arr1_.shape(); }
 };
@@ -71,7 +71,7 @@ class ArrayDiv : public ArrayExpr<ArrayDiv<A, B, Ndim>, Ndim> {
     ArrayDiv(const A &arr1, const B &arr2) : arr1_(arr1), arr2_(arr2) {
         assert(arr1.size() == arr2.size());
     }
-    auto operator[](int i) const { return arr1_[i] / arr2_[i]; }
+    auto operator[](size_t i) const { return arr1_[i] / arr2_[i]; }
     size_t size() const { return arr1_.size(); }
     std::array<ssize_t, Ndim> shape() const { return arr1_.shape(); }
 };
